Refactored Card Storage.  Added Purple Cards, Stances, and Powers.
