+++ src/main/java/Spansion/Cards/Blue/ArchaicFuel.java
@@ -1,9 +1,9 @@
-package Spansion.Cards;
+package Spansion.Cards.Blue;
 
+import Spansion.Powers.ArchaicFuelPower;
 import Spansion.Spansion;
 import basemod.abstracts.CustomCard;
 import com.megacrit.cardcrawl.actions.common.ApplyPowerAction;
-import com.megacrit.cardcrawl.cards.AbstractCard;
 import com.megacrit.cardcrawl.characters.AbstractPlayer;
 import com.megacrit.cardcrawl.core.CardCrawlGame;
 import com.megacrit.cardcrawl.dungeons.AbstractDungeon;
@@ -11,14 +11,13 @@ import com.megacrit.cardcrawl.localization.CardStrings;
 import com.megacrit.cardcrawl.monsters.AbstractMonster;
 
 import static Spansion.Spansion.makeCardPath;
-import Spansion.Powers.AspectOfTheCrowPower;
 
-public class AspectOfTheCrow extends CustomCard {
+public class ArchaicFuel extends CustomCard {
     // "Aspect of the Crow" - Power - Every time you play a card, gain 1 temporary strength and 1 temporary dexterity.
-    public static final String ID = Spansion.makeID(AspectOfTheCrow.class.getSimpleName());
+    public static final String ID = Spansion.makeID(ArchaicFuel.class.getSimpleName());
     private static final CardStrings cardStrings = CardCrawlGame.languagePack.getCardStrings(ID);
 
-    public static final String IMG = makeCardPath("Power.png");
+    public static final String IMG = makeCardPath("ArchaicFuel.png");
 
 
     public static final String NAME = cardStrings.NAME;
@@ -30,17 +29,17 @@ public class AspectOfTheCrow extends CustomCard {
 
     // STAT DECLARATION
 
-    private static final AbstractCard.CardRarity RARITY = CardRarity.COMMON;
-    private static final AbstractCard.CardTarget TARGET = CardTarget.SELF;
-    private static final AbstractCard.CardType TYPE = AbstractCard.CardType.POWER;
-    public static final AbstractCard.CardColor COLOR = AbstractCard.CardColor.COLORLESS;
+    private static final CardRarity RARITY = CardRarity.COMMON;
+    private static final CardTarget TARGET = CardTarget.SELF;
+    private static final CardType TYPE = CardType.POWER;
+    public static final CardColor COLOR = CardColor.COLORLESS;
 
     private static final int COST = 2;
 
     private static final int UPGRADE_PLUS_STACK = 1;
     private static final int UPGRADED_COST = 1;
 
-    public AspectOfTheCrow() {
+    public ArchaicFuel() {
         super(ID, NAME, IMG, COST, DESCRIPTION, TYPE, COLOR, RARITY, TARGET);
         magicNumber = baseMagicNumber = 1;
     }
@@ -49,7 +48,7 @@ public class AspectOfTheCrow extends CustomCard {
     @Override
     public void use(AbstractPlayer p, AbstractMonster m) {
         AbstractDungeon.actionManager.addToBottom(
-                new ApplyPowerAction(p, p, new AspectOfTheCrowPower(p,p, baseMagicNumber), baseMagicNumber)
+                new ApplyPowerAction(p, p, new ArchaicFuelPower(p,p, baseMagicNumber), baseMagicNumber)
         );
     }
