+++ src/main/java/Spansion/Cards/Colorless/OldOneTwo.java
@@ -1,0 +1,67 @@
+package Spansion.Cards.Colorless;
+
+import Spansion.Spansion;
+import basemod.abstracts.CustomCard;
+import com.megacrit.cardcrawl.actions.AbstractGameAction;
+import com.megacrit.cardcrawl.actions.common.ApplyPowerAction;
+import com.megacrit.cardcrawl.actions.common.DamageAction;
+import com.megacrit.cardcrawl.cards.DamageInfo;
+import com.megacrit.cardcrawl.characters.AbstractPlayer;
+import com.megacrit.cardcrawl.core.CardCrawlGame;
+import com.megacrit.cardcrawl.dungeons.AbstractDungeon;
+import com.megacrit.cardcrawl.localization.CardStrings;
+import com.megacrit.cardcrawl.monsters.AbstractMonster;
+
+import static Spansion.Spansion.makeCardPath;
+
+public class OldOneTwo extends CustomCard {
+//"The Old One-Two" - Attack - 1 - Deal 5 damage twice, applying 1 vulnerable each time.
+    public static final String ID = Spansion.makeID(OldOneTwo.class.getSimpleName());
+    private static final CardStrings cardStrings = CardCrawlGame.languagePack.getCardStrings(ID);
+
+    public static final String IMG = makeCardPath("Attack.png");
+
+    public static final String NAME = cardStrings.NAME;
+    public static final String DESCRIPTION = cardStrings.DESCRIPTION;
+
+    private static final CardRarity RARITY = CardRarity.COMMON;
+    private static final CardTarget TARGET = CardTarget.ENEMY;
+    private static final CardType TYPE = CardType.ATTACK;
+    public static final CardColor COLOR = CardColor.COLORLESS;
+
+    private static final int COST = 2;
+    private static final int DAMAGE = 4;
+    private static final int UPGRADE_PLUS_DMG = 2;
+    private static final int VULNERABLE_APPLIED = 1;
+
+    public OldOneTwo() {
+        super(ID, NAME, IMG, COST, DESCRIPTION, TYPE, COLOR, RARITY, TARGET);
+
+        damage = baseDamage = DAMAGE;
+        magicNumber = baseMagicNumber = VULNERABLE_APPLIED;
+    }
+
+    // Actions the card should do.
+    @Override
+    public void use(AbstractPlayer p, AbstractMonster m) {
+        for(int i = 0; i < 2; i++) {
+            AbstractDungeon.actionManager.addToBottom(
+                    new DamageAction(m, new DamageInfo(p, damage, damageTypeForTurn),
+                            AbstractGameAction.AttackEffect.SLASH_HORIZONTAL));
+            AbstractDungeon.actionManager.addToBottom(
+                    new ApplyPowerAction(m, p, new com.megacrit.cardcrawl.powers.VulnerablePower(p, baseMagicNumber, false)
+                    , baseMagicNumber)
+            );
+        }
+    }
+
+    // Upgraded stats.
+    @Override
+    public void upgrade() {
+        if (!upgraded) {
+            upgradeName();
+            upgradeDamage(UPGRADE_PLUS_DMG);
+            initializeDescription();
+        }
+    }
+}
