+++ src/main/java/Spansion/Cards/Green/FeintAttack.java
@@ -1,26 +1,28 @@
-package Spansion.Cards;
+package Spansion.Cards.Green;
 
 import Spansion.Spansion;
 import basemod.abstracts.CustomCard;
-import basemod.helpers.BaseModCardTags;
 import com.megacrit.cardcrawl.actions.AbstractGameAction;
 import com.megacrit.cardcrawl.actions.common.ApplyPowerAction;
 import com.megacrit.cardcrawl.actions.common.DamageAction;
+import com.megacrit.cardcrawl.actions.common.GainBlockAction;
+import com.megacrit.cardcrawl.actions.common.DiscardAction;
 import com.megacrit.cardcrawl.cards.DamageInfo;
 import com.megacrit.cardcrawl.characters.AbstractPlayer;
 import com.megacrit.cardcrawl.core.CardCrawlGame;
 import com.megacrit.cardcrawl.dungeons.AbstractDungeon;
 import com.megacrit.cardcrawl.localization.CardStrings;
 import com.megacrit.cardcrawl.monsters.AbstractMonster;
+import com.megacrit.cardcrawl.powers.NextTurnBlockPower;
 
 import static Spansion.Spansion.makeCardPath;
 
-public class OldOneTwo extends CustomCard {
-//"The Old One-Two" - Attack - 1 - Deal 5 damage twice, applying 1 vulnerable each time.
-    public static final String ID = Spansion.makeID(OldOneTwo.class.getSimpleName());
+public class FeintAttack extends CustomCard {
+    //"Feint Attack" - Attack - Discard a card.  Deal 10 damage.  Gain 4 block next turn.
+    public static final String ID = Spansion.makeID(FeintAttack.class.getSimpleName());
     private static final CardStrings cardStrings = CardCrawlGame.languagePack.getCardStrings(ID);
 
-    public static final String IMG = makeCardPath("Attack.png");
+    public static final String IMG = makeCardPath("FeintAttack.png");
 
     public static final String NAME = cardStrings.NAME;
     public static final String DESCRIPTION = cardStrings.DESCRIPTION;
@@ -28,32 +30,39 @@ public class OldOneTwo extends CustomCard {
     private static final CardRarity RARITY = CardRarity.COMMON;
     private static final CardTarget TARGET = CardTarget.ENEMY;
     private static final CardType TYPE = CardType.ATTACK;
-    public static final CardColor COLOR = CardColor.COLORLESS;
+    public static final CardColor COLOR = CardColor.GREEN;
 
     private static final int COST = 2;
-    private static final int DAMAGE = 4;
-    private static final int UPGRADE_PLUS_DMG = 2;
-    private static final int VULNERABLE_APPLIED = 1;
+    private static final int DAMAGE = 10;
+    private static final int UPGRADE_PLUS_DMG = 4;
+    private static final int BLOCK = 4;
+    private static final int UPGRADE_PLUS_BLOCK = 2;
 
-    public OldOneTwo() {
+    public FeintAttack() {
         super(ID, NAME, IMG, COST, DESCRIPTION, TYPE, COLOR, RARITY, TARGET);
 
         damage = baseDamage = DAMAGE;
-        magicNumber = baseMagicNumber = VULNERABLE_APPLIED;
+        block = baseBlock = BLOCK;
+
     }
 
     // Actions the card should do.
     @Override
     public void use(AbstractPlayer p, AbstractMonster m) {
-        for(int i = 0; i < 2; i++) {
-            AbstractDungeon.actionManager.addToBottom(
-                    new DamageAction(m, new DamageInfo(p, damage, damageTypeForTurn),
+        AbstractDungeon.actionManager.addToBottom(
+                new DiscardAction(p, p, 1, false)
+        );
+        AbstractDungeon.actionManager.addToBottom(
+                new DamageAction(m, new DamageInfo(p, damage, damageTypeForTurn),
                             AbstractGameAction.AttackEffect.SLASH_HORIZONTAL));
-            AbstractDungeon.actionManager.addToBottom(
-                    new ApplyPowerAction(m, p, new com.megacrit.cardcrawl.powers.VulnerablePower(p, baseMagicNumber, false)
-                    , baseMagicNumber)
+        AbstractDungeon.actionManager.addToBottom(
+                new GainBlockAction(p, p, block)
             );
-        }
+        AbstractDungeon.actionManager.addToBottom(
+                new ApplyPowerAction(p, p,
+                        new NextTurnBlockPower(p, block)
+                )
+        );
     }
 
     // Upgraded stats.
@@ -62,6 +71,7 @@ public class OldOneTwo extends CustomCard {
         if (!upgraded) {
             upgradeName();
             upgradeDamage(UPGRADE_PLUS_DMG);
+            upgradeBlock(UPGRADE_PLUS_BLOCK);
             initializeDescription();
         }
     }
