--- src/main/java/Spansion/Cards/LethalInjection.java
@@ -1,81 +1,0 @@
-package Spansion.Cards;
-
-import Spansion.Orbs.ToxicOrb;
-import Spansion.Spansion;
-import basemod.abstracts.CustomCard;
-import com.megacrit.cardcrawl.actions.AbstractGameAction;
-import com.megacrit.cardcrawl.actions.common.ApplyPowerAction;
-import com.megacrit.cardcrawl.actions.common.DamageAction;
-import com.megacrit.cardcrawl.actions.defect.ChannelAction;
-import com.megacrit.cardcrawl.cards.AbstractCard;
-import com.megacrit.cardcrawl.cards.DamageInfo;
-import com.megacrit.cardcrawl.characters.AbstractPlayer;
-import com.megacrit.cardcrawl.core.CardCrawlGame;
-import com.megacrit.cardcrawl.dungeons.AbstractDungeon;
-import com.megacrit.cardcrawl.localization.CardStrings;
-import com.megacrit.cardcrawl.monsters.AbstractMonster;
-import com.megacrit.cardcrawl.powers.PoisonPower;
-
-import static Spansion.Spansion.makeCardPath;
-
-public class LethalInjection extends CustomCard {
-    //"Lethal Injection" - Attack - 1 - Deal 3 damage.  Gain a Toxic Orb.  Increase this damage by 2 for every 0 cost card played this turn.
-    public static final String ID = Spansion.makeID(LethalInjection.class.getSimpleName());
-    private static final CardStrings cardStrings = CardCrawlGame.languagePack.getCardStrings(ID);
-
-    public static final String IMG = makeCardPath("LethalInjection.png");
-
-    public static final String NAME = cardStrings.NAME;
-    public static final String DESCRIPTION = cardStrings.DESCRIPTION;
-
-    private static final CardRarity RARITY = CardRarity.COMMON;
-    private static final CardTarget TARGET = CardTarget.ENEMY;
-    private static final CardType TYPE = CardType.ATTACK;
-    public static final CardColor COLOR = CardColor.BLUE;
-
-    private static final int COST = 1;
-    private static final int DAMAGE = 3;
-    private static final int UPGRADE_PLUS_DMG = 1;
-    private static final int BONUS_PER_O_PLAYED = 1;
-
-    private static int _zeroCostPlayed = 0;
-
-    public LethalInjection() {
-        super(ID, NAME, IMG, COST, DESCRIPTION, TYPE, COLOR, RARITY, TARGET);
-
-        damage = baseDamage = DAMAGE;
-        magicNumber = baseMagicNumber = BONUS_PER_O_PLAYED;
-    }
-
-    // Actions the card should do.
-    @Override
-    public void use(AbstractPlayer p, AbstractMonster m) {
-        AbstractDungeon.actionManager.addToBottom(
-                new DamageAction(m, new DamageInfo(p, damage, damageTypeForTurn),
-                        AbstractGameAction.AttackEffect.POISON));
-        AbstractDungeon.actionManager.addToBottom(
-                new ChannelAction(new ToxicOrb())
-        );
-    }
-
-    @Override
-    public void onPlayCard(AbstractCard c, AbstractMonster m) {
-        super.onPlayCard(c, m);
-        Spansion.logger.info("Played card: " + c.name);
-        if(c.cost == 0){
-            _zeroCostPlayed++;
-            Spansion.logger.info("Card is 0 cost.");
-            baseDamage += magicNumber;
-        }
-    }
-
-    // Upgraded stats.
-    @Override
-    public void upgrade() {
-        if (!upgraded) {
-            upgradeName();
-            upgradeDamage(UPGRADE_PLUS_DMG);
-            initializeDescription();
-        }
-    }
-}
