+++ src/main/java/Spansion/Cards/Purple/BalanceInAll.java
@@ -1,0 +1,47 @@
+package Spansion.Cards.Purple;
+
+import Spansion.Powers.BalanceInAllPower;
+import Spansion.Spansion;
+import basemod.abstracts.CustomCard;
+import com.megacrit.cardcrawl.actions.common.ApplyPowerAction;
+import com.megacrit.cardcrawl.characters.AbstractPlayer;
+import com.megacrit.cardcrawl.core.CardCrawlGame;
+import com.megacrit.cardcrawl.dungeons.AbstractDungeon;
+import com.megacrit.cardcrawl.localization.CardStrings;
+import com.megacrit.cardcrawl.monsters.AbstractMonster;
+
+import static Spansion.Spansion.makeCardPath;
+
+public class BalanceInAll extends CustomCard {
+
+    public static final String ID = Spansion.makeID(BalanceInAll.class.getSimpleName());
+    private static final CardStrings cardStrings = CardCrawlGame.languagePack.getCardStrings(ID);
+
+    public static final String IMG = makeCardPath("Power.png");
+    public static final String NAME = cardStrings.NAME;
+    public static final String DESCRIPTION = cardStrings.DESCRIPTION;
+
+    private static final CardRarity RARITY = CardRarity.UNCOMMON;
+    private static final CardTarget TARGET = CardTarget.SELF;
+    private static final CardType TYPE = CardType.POWER;
+    public static final CardColor COLOR = CardColor.PURPLE;
+
+    private static final int COST = 1;
+
+    public BalanceInAll() {
+        super(ID, NAME, IMG, COST, DESCRIPTION, TYPE, COLOR, RARITY, TARGET);
+    }
+
+    @Override
+    public void upgrade() {
+        if(!upgraded){
+            upgradeName();
+            initializeDescription();
+        }
+    }
+
+    @Override
+    public void use(AbstractPlayer abstractPlayer, AbstractMonster abstractMonster) {
+            AbstractDungeon.actionManager.addToBottom(new ApplyPowerAction(abstractPlayer, abstractPlayer, new BalanceInAllPower(abstractPlayer, abstractPlayer, 1)));
+    }
+}
