+++ src/main/java/Spansion/Cards/Red/CauterizingBlood.java
@@ -1,9 +1,10 @@
-package Spansion.Cards;
+package Spansion.Cards.Red;
 
-import Spansion.Powers.OnTheFlyPower;
+import Spansion.Powers.CauterizingBloodPower;
 import Spansion.Spansion;
 import basemod.abstracts.CustomCard;
 import com.megacrit.cardcrawl.actions.common.ApplyPowerAction;
+import com.megacrit.cardcrawl.cards.AbstractCard;
 import com.megacrit.cardcrawl.characters.AbstractPlayer;
 import com.megacrit.cardcrawl.core.CardCrawlGame;
 import com.megacrit.cardcrawl.dungeons.AbstractDungeon;
@@ -12,12 +13,13 @@ import com.megacrit.cardcrawl.monsters.AbstractMonster;
 
 import static Spansion.Spansion.makeCardPath;
 
-public class OnTheFly extends CustomCard {
+public class CauterizingBlood extends CustomCard {
+    // "Cauterizing Blood" - Power - 3 - At the end of your turn, heal for half of the damage you've taken this turn, rounded down.
 
-    public static final String ID = Spansion.makeID(OnTheFly.class.getSimpleName());
+    public static final String ID = Spansion.makeID(CauterizingBlood.class.getSimpleName());
     private static final CardStrings cardStrings = CardCrawlGame.languagePack.getCardStrings(ID);
 
-    public static final String IMG = makeCardPath("OnTheFly.png");
+    public static final String IMG = makeCardPath("CauterizingBlood.png");
 
 
     public static final String NAME = cardStrings.NAME;
@@ -29,25 +31,24 @@ public class OnTheFly extends CustomCard {
 
     // STAT DECLARATION
 
-    private static final CardRarity RARITY = CardRarity.COMMON;
-    private static final CardTarget TARGET = CardTarget.SELF;
-    private static final CardType TYPE = CardType.POWER;
-    public static final CardColor COLOR = CardColor.COLORLESS;
+    private static final AbstractCard.CardRarity RARITY = CardRarity.UNCOMMON;
+    private static final AbstractCard.CardTarget TARGET = CardTarget.SELF;
+    private static final AbstractCard.CardType TYPE = AbstractCard.CardType.POWER;
+    public static final AbstractCard.CardColor COLOR = AbstractCard.CardColor.RED;
 
     private static final int COST = 1;
+    private static final int UPGRADED_COST = 1;
 
-    private static final int UPGRADE_MINUS_COST = 1;
-
-    public OnTheFly() {
+    public CauterizingBlood() {
         super(ID, NAME, IMG, COST, DESCRIPTION, TYPE, COLOR, RARITY, TARGET);
-        magicNumber = baseMagicNumber = 1;
+        baseMagicNumber = magicNumber = 2;
     }
 
     // Actions the card should do.
     @Override
     public void use(AbstractPlayer p, AbstractMonster m) {
         AbstractDungeon.actionManager.addToBottom(
-                new ApplyPowerAction(p, p, new OnTheFlyPower(p, p, baseMagicNumber), baseMagicNumber)
+                new ApplyPowerAction(p, p, new CauterizingBloodPower(p, p, baseMagicNumber), baseMagicNumber)
         );
     }
 
@@ -56,8 +57,8 @@ public class OnTheFly extends CustomCard {
     public void upgrade() {
         if (!upgraded) {
             upgradeName();
-            upgradeBaseCost(0);
-            upgradeMagicNumber(UPGRADE_MINUS_COST);
+            upgradeBaseCost(UPGRADED_COST);
+            //upgradeMagicNumber(UPGRADE_MINUS_COST);
             rawDescription = UPGRADE_DESCRIPTION;
             initializeDescription();
         }
