+++ src/main/java/Spansion/Powers/StaggerPower.java
@@ -5,7 +5,10 @@ import Spansion.util.TextureLoader;
 import basemod.interfaces.CloneablePowerInterface;
 import com.badlogic.gdx.graphics.Texture;
 import com.badlogic.gdx.graphics.g2d.TextureAtlas;
-import com.megacrit.cardcrawl.actions.common.HealAction;
+import com.megacrit.cardcrawl.actions.AbstractGameAction;
+import com.megacrit.cardcrawl.actions.common.DamageAction;
+import com.megacrit.cardcrawl.actions.common.LoseHPAction;
+import com.megacrit.cardcrawl.cards.DamageInfo;
 import com.megacrit.cardcrawl.core.AbstractCreature;
 import com.megacrit.cardcrawl.core.CardCrawlGame;
 import com.megacrit.cardcrawl.dungeons.AbstractDungeon;
@@ -14,22 +17,20 @@ import com.megacrit.cardcrawl.powers.AbstractPower;
 
 import static Spansion.Spansion.makePowerPath;
 
-public class CauterizingBloodPower  extends AbstractPower implements CloneablePowerInterface {
+public class StaggerPower extends AbstractPower implements CloneablePowerInterface {
     public AbstractCreature source;
 
-    public static final String POWER_ID = Spansion.makeID(CauterizingBloodPower.class.getSimpleName());
+    public static final String POWER_ID = Spansion.makeID(StaggerPower.class.getSimpleName());
     private static final PowerStrings powerStrings = CardCrawlGame.languagePack.getPowerStrings(POWER_ID);
     public static final String NAME = powerStrings.NAME;
     public static final String[] DESCRIPTIONS = powerStrings.DESCRIPTIONS;
 
     // We create 2 new textures *Using This Specific Texture Loader* - an 84x84 image and a 32x32 one.
     // There's a fallback "missing texture" image, so the game shouldn't crash if you accidentally put a non-existent file.
-    private static final Texture tex84 = TextureLoader.getTexture(makePowerPath("placeholder_power84.png"));
-    private static final Texture tex32 = TextureLoader.getTexture(makePowerPath("placeholder_power32.png"));
+    private static final Texture tex84 = TextureLoader.getTexture(makePowerPath("archaicfuel_power84.png"));
+    private static final Texture tex32 = TextureLoader.getTexture(makePowerPath("archaicfuel_power32.png"));
 
-    private int damageTaken = 0;
-
-    public CauterizingBloodPower(final AbstractCreature owner, final AbstractCreature source, final int amount) {
+    public StaggerPower(final AbstractCreature owner, final AbstractCreature source, final int amount) {
         name = NAME;
         ID = POWER_ID;
 
@@ -46,35 +47,28 @@ public class CauterizingBloodPower  extends AbstractPower implements CloneablePo
         updateDescription();
     }
 
-    @Override
-    public void atStartOfTurn() {
-        damageTaken = 0;
-    }
 
     @Override
-    public int onLoseHp(int damageAmount) {
+    public void atEndOfTurn(boolean isPlayer) {
+        super.atEndOfTurn(isPlayer);
 
-        damageTaken+= damageAmount;
-        return super.onLoseHp(damageAmount);
-    }
+        int damage = Math.max(1, (int)Math.round(this.amount * .33));
+        AbstractDungeon.actionManager.addToBottom( new LoseHPAction(owner, owner, damage, AbstractGameAction.AttackEffect.SMASH));
+        this.amount = Math.max(0, amount - damage);
 
-    @Override
-    public void atEndOfTurn(final boolean isPlayer) {
-        int heal = (damageTaken + 1)/2;
-        if(heal > 0){
-            AbstractDungeon.actionManager.addToBottom(new HealAction(source, source, heal));
-        }
+        updateDescription();
     }
 
     // Update the description when you apply this power. (i.e. add or remove an "s" in keyword(s))
     @Override
     public void updateDescription() {
-        description = DESCRIPTIONS[0] + amount + DESCRIPTIONS[1];
+        int damage = Math.max(3, (int)Math.round(this.amount * .33));
+        description = DESCRIPTIONS[0] + damage + DESCRIPTIONS[1] + damage + DESCRIPTIONS[2];
 
     }
 
     @Override
     public AbstractPower makeCopy() {
-        return new CauterizingBloodPower(owner, source, amount);
+        return new StaggerPower(owner, source, amount);
     }
 }
