+++ src/main/resources/SpansionResources/localization/eng/Spansion-Card-Strings.json
@@ -48,8 +48,14 @@
     "NAME": "Archaic Fuel",
     "DESCRIPTION": "When you exhaust a card, channel 1 Toxic Orb."
   },
-  "Spansion:DrunkenSkill": {
-    "NAME": "Drunken Skill",
-    "DESCRIPTION": "Enter Drunken Stance."
+  "Spansion:CalculatedStrike": {
+    "NAME": "Calculated Strike",
+    "DESCRIPTION": "Deal !D! Damage.  Increases based on Your stagger and the enemy's Mark"
+  },"Spansion:AFortifyingDrink": {
+  "NAME": "A Fortifying Drink",
+  "DESCRIPTION": "Gain !B! armor if Drunken, otherwise enter Drunken Stance."
+  },"Spansion:BalanceInAll": {
+  "NAME": "Balance in all Things",
+  "DESCRIPTION": "At the end of your turn, gain block if in Drunken Stance, Draw next turn if Calm, and Exit Wrath if in Wrath.  Exiting Wrath Depletes this power."
   }
 }
